+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/models/ConvertToParticipantTimestampInfoModel.kt
@@ -1,0 +1,21 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.android.communication.ui.chat.models
+
+import org.threeten.bp.OffsetDateTime
+
+internal object ConvertToParticipantTimestampInfoModel {
+
+    fun fromMessageInfoModel(infoModel: MessageInfoModel) =
+        ParticipantTimestampInfoModel(
+            userIdentifier = infoModel.senderCommunicationIdentifier!!,
+            receivedOn = infoModel.createdOn!!
+        )
+
+    fun fromRemoteParticipantsInfoModel(infoModel: RemoteParticipantInfoModel) =
+        ParticipantTimestampInfoModel(
+            userIdentifier = infoModel.userIdentifier,
+            receivedOn = OffsetDateTime.now()
+        )
+}
