+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/models/ParticipantTimestampInfoModel.kt
@@ -9,18 +9,4 @@ import org.threeten.bp.OffsetDateTime
 internal data class ParticipantTimestampInfoModel(
     val userIdentifier: CommunicationIdentifier,
     val receivedOn: OffsetDateTime,
-) : BaseInfoModel {
-    companion object {
-        fun fromMessageInfoModel(infoModel: MessageInfoModel) =
-            ParticipantTimestampInfoModel(
-                userIdentifier = infoModel.senderCommunicationIdentifier!!,
-                receivedOn = infoModel.createdOn!!
-            )
-
-        fun fromRemoteParticipantsInfoModel(infoModel: RemoteParticipantInfoModel) =
-            ParticipantTimestampInfoModel(
-                userIdentifier = infoModel.userIdentifier,
-                receivedOn = OffsetDateTime.now()
-            )
-    }
-}
+) : BaseInfoModel
